added route to import csv files

+++ src/database/migrations/1605286983383-CreateTransactions.ts
@@ -24,7 +24,7 @@ export default class CreateTransactions1605286983383
           },
           {
             name: 'value',
-            type: 'float',
+            type: 'int',
             isNullable: false,
           },
           {
